+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 2.0.0
+
+- Added a timeout parameter to the `generateBoard` function. All the parameters of the function are now named and optional.
+- Removed the artificial limit on the number of empty positions of generated puzzles with dimensions 16 or 25.
+
+Motivated by https://github.com/raulcostajunior/fludoku/issues/1 (thanks @ahmedatef286 for reporting!)
+
 ## 1.0.0
 
 - Initial version with full capabilities (generator lib completed)